清理告警

+++ device/plugins/native_hook/src/hook_client.cpp
@@ -986,7 +986,7 @@ int  ohos_malloc_hook_prctl(int option, unsigned long arg2, unsigned long arg3,
 
 bool ohos_set_filter_size(size_t size, void* ret)
 {
-    if (g_ClientConfig.filterSize < 0 || size < g_ClientConfig.filterSize || size > g_maxSize) {
+    if (g_ClientConfig.filterSize < 0 || size < static_cast<size_t>(g_ClientConfig.filterSize) || size > g_maxSize) {
         return false;
     }
     return true;
